introduce preparereaders
since we use lazyreaders internally we only request data when we
actually need it. that has the disadvantage of incuring a lot of
overhead waiting for the transfer to begin. the preparereaders help
offset that by starting subsequent readers ahead of time. at the cost of
memory consumption of course.